+++ src/shared/io_system/io_vtk_mesh.cpp
@@ -7,9 +7,126 @@ BodyStatesRecordingToMeshVtp::BodyStatesRecordingToMeshVtp(SPHBody &body, ANSYSM
     : BodyStatesRecordingToVtp(body), node_coordinates_(ansys_mesh.node_coordinates_),
       elements_nodes_connection_(ansys_mesh.elements_nodes_connection_) {}
 //=================================================================================================//
+void BodyStatesRecordingToMeshVtp::writeWithFileName(const std::string &sequence)
+{
+    for (SPHBody *body : bodies_)
+    {
+        if (body->checkNewlyUpdated() && state_recording_)
+        {
+            // TODO: we can short the file name by without using SPHBody
+            std::string filefullpath = io_environment_.output_folder_ + "/" + body->getName() + "_" + sequence + ".vtp";
+            if (fs::exists(filefullpath))
+            {
+                fs::remove(filefullpath);
+            }
+            std::ofstream out_file(filefullpath.c_str(), std::ios::trunc);
+            // begin of the XML file
+            out_file << "<?xml version=\"1.0\"?>\n";
+            out_file << "<VTKFile type=\"PolyData\" version=\"1.0\" byte_order=\"LittleEndian\">\n";
+            out_file << "<PolyData>\n";
+
+            // Write point data
+            out_file << "<Piece NumberOfPoints=\"" << node_coordinates_.size()
+                     << "\" NumberOfVerts=\"0\" NumberOfLines=\"0\" NumberOfStrips=\"0\" NumberOfPolys=\""
+                     << elements_nodes_connection_.size() << "\">\n";
+            out_file << "<Points>\n";
+            out_file << "<DataArray type=\"Float64\" NumberOfComponents=\"3\" format=\"ascii\">\n";
+
+            size_t total_nodes = node_coordinates_.size();
+            for (size_t node = 0; node != total_nodes; ++node)
+            {
+                Vec3d particle_position = upgradeToVec3d(node_coordinates_[node]);
+                out_file << particle_position[0] << " " << particle_position[1] << " " << particle_position[2] << "\n";
+            }
+
+            out_file << "</DataArray>\n";
+            out_file << "</Points>\n";
+
+            // Write face data
+            out_file << "<Polys>\n";
+            out_file << "<DataArray type=\"Int32\" Name=\"connectivity\" format=\"ascii\">\n";
+
+            for (const auto &element : elements_nodes_connection_)
+            {
+                for (const auto &vertex : element)
+                {
+                    out_file << vertex << " ";
+                }
+                out_file << "\n";
+            }
+
+            out_file << "</DataArray>\n";
+            out_file << "<DataArray type=\"Int32\" Name=\"offsets\" format=\"ascii\">\n";
+
+            size_t offset = 0;
+            for (const auto &face : elements_nodes_connection_)
+            {
+                offset += face.size();
+                out_file << offset << " ";
+            }
+
+            out_file << "\n</DataArray>\n";
+            out_file << "</Polys>\n";
+
+            // Write face attribute data
+            out_file << "<CellData>\n";
+
+            BaseParticles &particles = body->getBaseParticles();
+            writeParticlesToVtk(out_file, particles);
+
+            out_file << "</CellData>\n";
+
+            // Write file footer
+            out_file << "</Piece>\n";
+            out_file << "</PolyData>\n";
+            out_file << "</VTKFile>\n";
+
+            out_file.close();
+        }
+        body->setNotNewlyUpdated();
+    }
+}
+//=================================================================================================//
 BodyStatesRecordingToMeshVtu::BodyStatesRecordingToMeshVtu(SPHBody &body, ANSYSMesh &ansys_mesh)
     : BodyStatesRecordingToVtp(body), node_coordinates_(ansys_mesh.node_coordinates_),
       elements_nodes_connection_(ansys_mesh.elements_nodes_connection_), bounds_(body) {};
+//=================================================================================================//
+void BodyStatesRecordingToMeshVtu::writeWithFileName(const std::string &sequence)
+{
+    for (SPHBody *body : bodies_)
+    {
+        if (body->checkNewlyUpdated() && state_recording_)
+        {
+            std::string filefullpath = io_environment_.output_folder_ + "/SPHBody_" + body->getName() + "_" + sequence + ".vtu";
+            if (fs::exists(filefullpath))
+            {
+                fs::remove(filefullpath);
+            }
+            std::ofstream out_file(filefullpath.c_str(), std::ios::trunc);
+
+            FileHeader(out_file);
+            Real range_max = FileNodeCoordinates(out_file);
+            FileInformationKey(out_file, range_max);
+            FileCellConnectivity(out_file);
+            FileOffsets(out_file);
+            FileTypeOfCell(out_file);
+
+            // write Particle data to vtu file
+            out_file << "<CellData>\n";
+
+            BaseParticles &particles = body->getBaseParticles();
+            writeParticlesToVtk(out_file, particles);
+
+            out_file << "</CellData>\n";
+            // Write VTU file footer
+            out_file << "</Piece>\n";
+            out_file << "</UnstructuredGrid>\n";
+            out_file << "</VTKFile>\n";
+            out_file.close();
+        }
+        body->setNotNewlyUpdated();
+    }
+}
 //=================================================================================================//
 void BodyStatesRecordingToMeshVtu::FileHeader(std::ofstream &out_file)
 {
@@ -22,12 +139,11 @@ void BodyStatesRecordingToMeshVtu::FileHeader(std::ofstream &out_file)
     out_file << "</FieldData>\n";
 }
 //=================================================================================================//
-void BodyStatesRecordingToMeshVtu::FileNodeCoordinates(
-    std::ofstream &out_file, StdLargeVec<Vecd> &node_coordinates_,
-    StdLargeVec<StdVec<size_t>> &elements_nodes_connection_, SPHBody &bounds_, Real &rangemax)
+Real BodyStatesRecordingToMeshVtu::FileNodeCoordinates(std::ofstream &out_file)
 {
     // Write point data
-    out_file << "<Piece NumberOfPoints=\"" << node_coordinates_.size() << "\" NumberOfCells=\"" << elements_nodes_connection_.size() << "\">\n";
+    out_file << "<Piece NumberOfPoints=\"" << node_coordinates_.size() << "\" NumberOfCells=\""
+             << elements_nodes_connection_.size() << "\">\n";
     out_file << "<PointData>\n";
     out_file << "</PointData>\n";
     out_file << "<CellData>\n";
@@ -36,14 +152,16 @@ void BodyStatesRecordingToMeshVtu::FileNodeCoordinates(
     BoundingBox bounds = bounds_.getSPHSystemBounds();
     Real first_max = bounds.first_.cwiseAbs().maxCoeff();
     Real second_max = bounds.second_.cwiseAbs().maxCoeff();
-    rangemax = 1.03075 * (std::max(first_max, second_max));
-    out_file << "<DataArray type=\"Float64\" Name=\"Points\" NumberOfComponents=\"3\" format=\"ascii\" RangeMin=\"0\" RangeMax=\"" << rangemax << "\">\n";
+    Real rangemax = 1.03075 * (std::max(first_max, second_max));
+    out_file << "<DataArray type=\"Float64\" Name=\"Points\" NumberOfComponents=\"3\""
+             << " format=\"ascii\" RangeMin=\"0\" RangeMax=\"" << rangemax << "\">\n";
 
-    size_t total_nodes = node_coordinates_.size();
-    for (size_t node = 0; node != total_nodes; ++node)
+    for (size_t node = 0; node != node_coordinates_.size(); ++node)
     {
-        out_file << node_coordinates_[node][0] << " " << node_coordinates_[node][1] << " " << node_coordinates_[node][2] << " \n";
+        Vec3d node_position = upgradeToVec3d(node_coordinates_[node]);
+        out_file << node_position[0] << " " << node_position[1] << " " << node_position[2] << " \n";
     }
+    return rangemax;
 }
 //=================================================================================================//
 void BodyStatesRecordingToMeshVtu::FileInformationKey(std::ofstream &out_file, Real &rangemax)
@@ -68,13 +186,13 @@ void BodyStatesRecordingToMeshVtu::FileInformationKey(std::ofstream &out_file, R
     out_file << "</Points>\n";
 }
 //=================================================================================================//
-void BodyStatesRecordingToMeshVtu::FileCellConnectivity(
-    std::ofstream &out_file,
-    StdLargeVec<StdVec<size_t>> &elements_nodes_connection_, StdLargeVec<Vecd> &node_coordinates_)
+void BodyStatesRecordingToMeshVtu::FileCellConnectivity(std::ofstream &out_file)
 {
     out_file << "<Cells>\n";
     // Write Cell data
-    out_file << "<DataArray type=\"Int64\" Name=\"connectivity\" format=\"ascii\" RangeMin=\"0\" RangeMax=\"" << node_coordinates_.size() - 1 << "\">\n";
+    out_file << "<DataArray type=\"Int64\" Name=\"connectivity\""
+             << " format =\"ascii\" RangeMin=\"0\" RangeMax=\"" << node_coordinates_.size() - 1
+             << "\">\n";
 
     for (const auto &cell : elements_nodes_connection_)
     {
@@ -88,10 +206,11 @@ void BodyStatesRecordingToMeshVtu::FileCellConnectivity(
     out_file << "</DataArray>\n";
 }
 //=================================================================================================//
-void BodyStatesRecordingToMeshVtu::FileOffsets(std::ofstream &out_file,
-                                               StdLargeVec<StdVec<size_t>> &elements_nodes_connection_)
+void BodyStatesRecordingToMeshVtu::FileOffsets(std::ofstream &out_file)
 {
-    out_file << "<DataArray type=\"Int64\" Name=\"offsets\" format=\"ascii\" RangeMin=\"4\" RangeMax=\"" << 4 * elements_nodes_connection_.size() << "\">\n";
+    out_file << "<DataArray type=\"Int64\" Name=\"offsets\""
+             << " format =\"ascii\" RangeMin=\"4\" RangeMax=\"" << 4 * elements_nodes_connection_.size()
+             << "\">\n";
 
     size_t offset = 0;
     for (const auto &face : elements_nodes_connection_)
@@ -102,8 +221,7 @@ void BodyStatesRecordingToMeshVtu::FileOffsets(std::ofstream &out_file,
     out_file << "\n</DataArray>\n";
 }
 //=================================================================================================//
-void BodyStatesRecordingToMeshVtu::FileTypeOfCell(std::ofstream &out_file,
-                                                  StdLargeVec<StdVec<size_t>> &elements_nodes_connection_)
+void BodyStatesRecordingToMeshVtu::FileTypeOfCell(std::ofstream &out_file)
 {
     size_t type = 10;
     // Specifies type of cell 10 = tetrahedral
